use wagtail-2.0 branch of wagtail/bakerydemo
This change updates this project to use a wagtail-2.0 branch of
wagtail/bakerydemo so that breaking changes in 2.0 can be easily worked
on without requiring an upgrade of bakerydemo.

See https://github.com/wagtail/wagtail/issues/3918 for background.

This change also adds `set -e` to the bash script so that it aborts if
any command fails.